Create ja-negative-sentiment
This is to capture negative responses on an individual's Facebook, Twitter, Instagram and Telegram accounts. The feedback is based on posts, shares, likes, retweets and comments made on the various social media platforms for the periods leading to/just after the Kenyan national elections i.e. June 2007 to June 2008; and June 2013 to June 2014; and January 2017 to June 2017.